@@ -20,27 +20,42 @@ plugins {
     // version of spring boot plugin is also resolved by 'settings.gradle.kts'
     id("org.springframework.boot")
     // other plugins require a version to be mentioned
-    id("io.spring.dependency-management") version "1.0.8.RELEASE"
-    id("org.owasp.dependencycheck") version "5.0.0"
-    id("io.gitlab.arturbosch.detekt") version "1.0.0-RC14"
+    id("io.spring.dependency-management") version "1.0.11.RELEASE"
+    id("io.gitlab.arturbosch.detekt") version "1.17.0"
+    id("com.github.ben-manes.versions") version "0.39.0"
 }
 
 configure<JavaPluginConvention> {
-    sourceCompatibility = JavaVersion.VERSION_1_8
-    targetCompatibility = JavaVersion.VERSION_1_8
+    sourceCompatibility = JavaVersion.VERSION_11
+    targetCompatibility = JavaVersion.VERSION_11
+}
+
+jacoco {
+    toolVersion = "0.8.7"
+}
+
+detekt {
+    buildUponDefaultConfig = true
+    config = files("src/main/resources/detekt-config.yml")
+    reports {
+        html.enabled = true
+        xml.enabled = true
+        txt.enabled = true
+        sarif.enabled = true
+    }
 }
 
 tasks {
     withType<KotlinCompile> {
         kotlinOptions {
-            jvmTarget = "1.8"
+            jvmTarget = "11"
             freeCompilerArgs = listOf("-Xjsr305=strict")
         }
     }
     withType<Test> {
         useJUnitPlatform {
-            systemProperty("spring.datasource.url", "jdbc:postgresql://localhost:45432/kotlink")
-            systemProperty("spring.redis.url", "redis://localhost:46379")
+            systemProperty("spring.datasource.url", "jdbc:postgresql://localhost:55432/kotlink")
+            systemProperty("spring.redis.url", "redis://localhost:6379")
         }
         testLogging.apply {
             events("passed", "skipped", "failed")
