@@ -4,7 +4,7 @@ import org.amshove.kluent.shouldBe
 import org.amshove.kluent.shouldBeGreaterThan
 import org.amshove.kluent.shouldContain
 import org.amshove.kluent.shouldEndWith
-import org.amshove.kluent.shouldEqual
+import org.amshove.kluent.shouldBeEqualTo
 import org.amshove.kluent.shouldNotBe
 import org.amshove.kluent.shouldStartWith
 import org.amshove.kluent.shouldThrow
@@ -61,21 +61,21 @@ class AliasRepoImplTest(
     @Test
     fun `'findById' should return an alias if provided ID matches one`() {
         repo.findById(testAlias.id).also {
-            it?.link shouldEqual testAlias.link
+            it?.link shouldBeEqualTo testAlias.link
         }
     }
 
     @Test
     fun `'findById' should return no alias if provided ID does not match any of the aliases`() {
         repo.findById(Long.MAX_VALUE).also {
-            it shouldEqual null
+            it shouldBeEqualTo null
         }
     }
 
     @Test
     fun `'findByFullLink' should return an alias if provided full link matches an alias`() {
         repo.findByFullLink(testAlias.fullLink).also {
-            it?.link shouldEqual testAlias.link
+            it?.link shouldBeEqualTo testAlias.link
         }
     }
 
@@ -89,7 +89,7 @@ class AliasRepoImplTest(
     @Test
     fun `'findByNamespace' should return not aliases if given keyword does not match any namespace`() {
         repo.findByNamespace(UUID.randomUUID().toString()).also {
-            it.size shouldEqual 0
+            it.size shouldBeEqualTo 0
         }
     }
 
@@ -103,7 +103,7 @@ class AliasRepoImplTest(
     @Test
     fun `'findByNamespaceAndLink' should return the alias matching the given link and namespace`() {
         repo.findByNamespaceAndLink(testNamespace.keyword, testAlias.link).also {
-            it?.link shouldEqual testAlias.link
+            it?.link shouldBeEqualTo testAlias.link
         }
     }
 
@@ -186,7 +186,7 @@ class AliasRepoImplTest(
         repo.countWithAllOfTermsInFullLink(
             terms = testAlias.link.split(" ").toList()
         ).also {
-            it shouldEqual 1
+            it shouldBeEqualTo 1
         }
     }
 
@@ -203,7 +203,7 @@ class AliasRepoImplTest(
             )
         ).also {
             it.id shouldBeGreaterThan 0
-            repo.findById(it.id)?.id shouldEqual it.id
+            repo.findById(it.id)?.id shouldBeEqualTo it.id
         }
     }
 
@@ -222,7 +222,7 @@ class AliasRepoImplTest(
                 redirectUrl = testAlias.redirectUrl + "abc"
             )
         ).also {
-            it.id shouldEqual testAlias.id
+            it.id shouldBeEqualTo testAlias.id
             repo.findById(it.id)?.link!! shouldEndWith "abc"
             repo.findById(it.id)?.redirectUrl!! shouldEndWith "abc"
         }
@@ -231,14 +231,14 @@ class AliasRepoImplTest(
     @Test
     fun `'deleteById' should return True if alias was found and deleted`() {
         repo.deleteById(testAlias.id).also {
-            it shouldEqual true
+            it shouldBeEqualTo true
         }
     }
 
     @Test
     fun `'deleteById' should return False if alias was not found`() {
         repo.deleteById(Long.MAX_VALUE).also {
-            it shouldEqual false
+            it shouldBeEqualTo false
         }
     }
 
