@@ -1,7 +1,7 @@
 package org.kotlink.ui
 
 import com.nhaarman.mockitokotlin2.whenever
-import org.amshove.kluent.shouldEqual
+import org.amshove.kluent.shouldBeEqualTo
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
 import org.mockito.Mock
@@ -21,6 +21,6 @@ class ViewUtilsTest {
         whenever(request.scheme).thenReturn("http")
         whenever(request.serverName).thenReturn("localhost")
 
-        viewUtils.serverUrlFromRequest(request) shouldEqual "http://localhost:8080/"
+        viewUtils.serverUrlFromRequest(request) shouldBeEqualTo "http://localhost:8080/"
     }
 }
