@@ -3,23 +3,19 @@ package org.kotlink.ui.search
 import com.nhaarman.mockitokotlin2.whenever
 import org.hamcrest.Matchers
 import org.junit.jupiter.api.Test
-import org.junit.jupiter.api.extension.ExtendWith
 import org.kotlink.INBOX_ALIAS
+import org.kotlink.WithMockMvcSetUp
 import org.kotlink.core.alias.AliasService
 import org.kotlink.perform
-import org.kotlink.ui.UiTestConfig
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest
 import org.springframework.boot.test.mock.mockito.MockBean
-import org.springframework.context.annotation.Import
-import org.springframework.test.context.junit.jupiter.SpringExtension
 import org.springframework.test.web.servlet.MockMvc
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders
 import org.springframework.test.web.servlet.result.MockMvcResultMatchers
 
-@ExtendWith(SpringExtension::class)
-@WebMvcTest(LinkSearchUiController::class, secure = false)
-@Import(UiTestConfig::class)
+@WebMvcTest(LinkSearchUiController::class)
+@WithMockMvcSetUp
 class LinkSearchUiControllerTest(
     @Autowired private val mvc: MockMvc
 ) {
