@@ -4,24 +4,20 @@ import com.nhaarman.mockitokotlin2.any
 import com.nhaarman.mockitokotlin2.whenever
 import org.hamcrest.Matchers
 import org.junit.jupiter.api.Test
-import org.junit.jupiter.api.extension.ExtendWith
 import org.kotlink.TEST_SECRET
+import org.kotlink.WithMockMvcSetUp
 import org.kotlink.core.secret.ApiSecretService
 import org.kotlink.perform
-import org.kotlink.ui.UiTestConfig
 import org.kotlink.ui.help.SetupInstructionsUiController
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest
 import org.springframework.boot.test.mock.mockito.MockBean
-import org.springframework.context.annotation.Import
-import org.springframework.test.context.junit.jupiter.SpringExtension
 import org.springframework.test.web.servlet.MockMvc
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders
 import org.springframework.test.web.servlet.result.MockMvcResultMatchers
 
-@ExtendWith(SpringExtension::class)
-@WebMvcTest(SetupInstructionsUiController::class, secure = false)
-@Import(UiTestConfig::class)
+@WebMvcTest(SetupInstructionsUiController::class)
+@WithMockMvcSetUp
 class SetupInstructionsUiControllerTest(
     @Autowired private val mvc: MockMvc
 ) {
